+++ src/btstack_event.h
@@ -2327,6 +2327,15 @@ static inline uint16_t bnep_event_channel_opened_get_mtu(const uint8_t * event){
 static inline void bnep_event_channel_opened_get_remote_address(const uint8_t * event, bd_addr_t remote_address){
     reverse_bd_addr(&event[11], remote_address);
 }
+/**
+ * @brief Get field con_handle from event BNEP_EVENT_CHANNEL_OPENED
+ * @param event packet
+ * @return con_handle
+ * @note: btstack_type H
+ */
+static inline hci_con_handle_t bnep_event_channel_opened_get_con_handle(const uint8_t * event){
+    return little_endian_read_16(event, 17);
+}
 
 /**
  * @brief Get field bnep_cid from event BNEP_EVENT_CHANNEL_CLOSED