+++ src/classic/bnep.c
@@ -79,7 +79,7 @@ static void bnep_emit_open_channel_complete(bnep_channel_t *channel, uint8_t sta
     log_info("BNEP_EVENT_CHANNEL_OPENED status 0x%02x bd_addr: %s, handler %p", status, bd_addr_to_str(channel->remote_addr), channel->packet_handler);
     if (!channel->packet_handler) return;
 
-    uint8_t event[3 + sizeof(bd_addr_t) + 4 * sizeof(uint16_t)];
+    uint8_t event[3 + sizeof(bd_addr_t) + 4 * sizeof(uint16_t) + 2];
     event[0] = BNEP_EVENT_CHANNEL_OPENED;
     event[1] = sizeof(event) - 2;
     event[2] = status;
@@ -88,6 +88,7 @@ static void bnep_emit_open_channel_complete(bnep_channel_t *channel, uint8_t sta
     little_endian_store_16(event, 7, channel->uuid_dest);
     little_endian_store_16(event, 9, channel->max_frame_size);
     reverse_bd_addr(channel->remote_addr, &event[11]);
+    little_endian_store_16(event, 17, channel->con_handle);
     hci_dump_packet( HCI_EVENT_PACKET, 0, event, sizeof(event));
 	(*channel->packet_handler)(HCI_EVENT_PACKET, 0, (uint8_t *) event, sizeof(event));
 }